refactor: reorganized directories
