+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
-import TodoList from './TodoList';
-import TodoForm from './TodoForm'
+import TodoList from './features/TodoList/TodoList';
+import TodoForm from './features/TodoForm';
 
 function App() {
